// easytree-final-backend-main/src/Enum.ts
export enum BLOCKSTATE {
    scheduled = "scheduled",
    onGoing = "onGoing",
    completed = "completed"
}

export enum BLOCKRESULT {
    winner = "w",
    loser = "l",
    notDecided = "nd"
}

export enum TOURNAMENTSTATE {
    signUpPhase = "signUpPhase",
    drawPhase = "drawPhase",
    onGoing = "onGoing",
    completed = "completed",
    discarded = "discarded"
}